First commit
A basic project structure has been introduced to make this
program modular and easy to add to. Format types are
implemented as their own respective plugins, and soon the
output mechanisms will implemented similiarly.

The interactive portion of the program will be implemented
as abstractly as possible, maybe handed off to it's own
driver so different types of play styles can be implemented.