Implemented formattypes, engines, and main
This is the first go at implementing a modular pipeline.
The formatter parses a text file of a specific format, then
passes it to the engine. The engine then modifies the data
according to how it was implemented. For example, the "fill
in the blank" engine creates randomized blanks on the ``body``
sectino of the returned result from the formatter. For now,
only one runner has been implemented - the command line runner.
The command line runner runs an interative prompt on the
command line, asking the user to write out the line with the
hollowed lines filled in.

Some example format type files have been provided