Add user-defined difficulty level
The difficulty level is equal to the number of words that are
replaced in a line. This commit adds the argument "--level"
to the command line, which allows the user to define the
number of words that are replaced in the line.

It also prepares the setup.py for the 0.0.1a4 release.