Fix bug when giving a high level
When the level was set higher then the number of words
in the line, the engine would get stuck in a loop.

This commit adds logic to consider the number of words
possible to replace, setting it initially to the number
of words in a line.

If the user-defined 'level' is higher then the number
of words possible to replace, then the engine just replaces
all the words in the line with underscores.

Fixes:
https://github.com/alextricity25/dwell_in_you_richly/issues/6