Added 'off' parameter for identifier position
Specifying the 'off' value for the 'identifier_position' key will tell
the runner to ignore the identifier. This might be useful in situations where
the identifier might not want to be printed. i.e. when the identifier is a
verse reference, and the user would want to be tested on the verse reference