Now consider leading whitespace
The outline format now considers leading whitespace
when printing each line. This does not affect what
the expected input is. In other words, the user does
not have to input the leading white space in order
to get a correct answer. They only have to input
the content of the line.

Additionally, this commits adds one space between the identifier
and the rest of the line. This is for readability. This does
change the expected user input. The user is now required to
also include this space in order to get a correct answer.